Complete setting up ts-jest
