+++ .jest.json
@@ -9,5 +9,6 @@
     "js",
     "jsx",
     "json"
-  ]
+  ],
+  "moduleDirectories": ["src", "node_modules"]
 }
